+++ lib/mongoid/persistable/atomic.rb
@@ -1,7 +1,6 @@
 # encoding: utf-8
 require "mongoid/persistable/atomic/operation"
 require "mongoid/persistable/atomic/add_to_set"
-require "mongoid/persistable/atomic/pull_all"
 require "mongoid/persistable/atomic/push"
 require "mongoid/persistable/atomic/push_all"
 
@@ -32,24 +31,6 @@ module Mongoid
       # @since 2.0.0
       def add_to_set(field, value, options = {})
         AddToSet.new(self, field, value, options).persist
-      end
-
-      # Performs an atomic $pullAll of the provided value on the supplied
-      # field. If the field does not exist it will be initialized as an
-      # empty array.
-      #
-      # @example Pull the values from the field.
-      #   person.pull_all(:aliases, [ "Bond", "James" ])
-      #
-      # @param [ Symbol ] field The name of the field.
-      # @param [ Array<Object> ] value The values to pull.
-      # @param [ Hash ] options The mongo persistence options.
-      #
-      # @return [ Array<Object> ] The new value of the field.
-      #
-      # @since 2.0.0
-      def pull_all(field, value, options = {})
-        PullAll.new(self, field, value, options).persist
       end
 
       # Performs an atomic $push of the provided value on the supplied field. If