+++ lib/mongoid/persistable/pullable.rb
@@ -32,8 +32,28 @@ module Mongoid
         end
       end
 
-      # def pull_all(pulls)
-      # end
+      # Pull multiple values from the provided array fields.
+      #
+      # @example Pull values from the arrays.
+      #   document.pull_all(names: [ "Jeff", "Bob" ], levels: [ 5, 6 ])
+      #
+      # @param [ Hash ] pulls The pull all operations.
+      #
+      # @return [ true, false ] If the operation succeeded.
+      #
+      # @since 4.0.0
+      def pull_all(pulls)
+        prepare_atomic_operation do |coll, selector, ops|
+          pulls.each do |field, value|
+            normalized = database_field_name(field)
+            existing = send(field) || []
+            value.each{ |val| existing.delete(val) }
+            remove_change(normalized)
+            ops[atomic_attribute_name(normalized)] = value
+          end
+          coll.find(selector).update(positionally(selector, "$pullAll" => ops))
+        end
+      end
     end
   end
 end