+++ spec/app/models/person.rb
@@ -31,6 +31,7 @@ class Person
   field :i, as: :inte, type: Integer
   field :a, as: :array, type: Array
   field :desc, localize: true
+  field :test_array, type: Array
 
   index age: 1
   index addresses: 1